+++ src/interpreter.rs
@@ -7,10 +7,17 @@ use ripemd::Ripemd160;
 use sha1::Sha1;
 use sha2::{Digest, Sha256};
 
-use super::external::pubkey::PubKey;
-use super::script::{Operation::*, PushValue::*, *};
-use super::script_error::*;
-use crate::signature;
+use crate::{
+    external::pubkey::PubKey,
+    script::{
+        parse_num, serialize_num, Opcode,
+        Operation::*,
+        PushValue::{self, *},
+        Script, LOCKTIME_THRESHOLD, MAX_SCRIPT_ELEMENT_SIZE, MAX_SCRIPT_SIZE,
+    },
+    script_error::ScriptError,
+    signature,
+};
 
 bitflags::bitflags! {
     #[derive(Copy, Clone, Debug, PartialEq, Eq)]
@@ -119,7 +126,7 @@ pub struct Stack<T>(Vec<T>);
 /// Wraps a Vec (or whatever underlying implementation we choose in a way that matches the C++ impl
 /// and provides us some decent chaining)
 impl<T: Clone> Stack<T> {
-    pub fn new() -> Self {
+    fn new() -> Self {
         Stack(vec![])
     }
 
@@ -132,58 +139,58 @@ impl<T: Clone> Stack<T> {
         }
     }
 
-    pub fn rget(&self, i: usize) -> Result<&T, ScriptError> {
+    fn rget(&self, i: usize) -> Result<&T, ScriptError> {
         let idx = self.rindex(i)?;
         self.0.get(idx).ok_or(ScriptError::InvalidStackOperation)
     }
 
-    pub fn rswap(&mut self, a: usize, b: usize) -> Result<(), ScriptError> {
+    fn rswap(&mut self, a: usize, b: usize) -> Result<(), ScriptError> {
         let ra = self.rindex(a)?;
         let rb = self.rindex(b)?;
         self.0.swap(ra, rb);
         Ok(())
     }
 
-    pub fn pop(&mut self) -> Result<T, ScriptError> {
+    fn pop(&mut self) -> Result<T, ScriptError> {
         self.0.pop().ok_or(ScriptError::InvalidStackOperation)
     }
 
-    pub fn push(&mut self, value: T) {
+    fn push(&mut self, value: T) {
         self.0.push(value)
     }
 
-    pub fn is_empty(&self) -> bool {
+    pub(crate) fn is_empty(&self) -> bool {
         self.0.is_empty()
     }
 
-    pub fn len(&self) -> usize {
+    pub(crate) fn len(&self) -> usize {
         self.0.len()
     }
 
-    pub fn iter(&self) -> Iter<'_, T> {
+    fn iter(&self) -> Iter<'_, T> {
         self.0.iter()
     }
 
-    pub fn last_mut(&mut self) -> Result<&mut T, ScriptError> {
+    fn last_mut(&mut self) -> Result<&mut T, ScriptError> {
         self.0.last_mut().ok_or(ScriptError::InvalidStackOperation)
     }
 
-    pub fn last(&self) -> Result<&T, ScriptError> {
+    fn last(&self) -> Result<&T, ScriptError> {
         self.0.last().ok_or(ScriptError::InvalidStackOperation)
     }
 
-    pub fn split_last(&self) -> Result<(&T, Stack<T>), ScriptError> {
+    fn split_last(&self) -> Result<(&T, Stack<T>), ScriptError> {
         self.0
             .split_last()
             .ok_or(ScriptError::InvalidStackOperation)
             .map(|(last, rem)| (last, Stack(rem.to_vec())))
     }
 
-    pub fn rremove(&mut self, start: usize) -> Result<T, ScriptError> {
+    fn rremove(&mut self, start: usize) -> Result<T, ScriptError> {
         self.rindex(start).map(|rstart| self.0.remove(rstart))
     }
 
-    pub fn rinsert(&mut self, i: usize, element: T) -> Result<(), ScriptError> {
+    fn rinsert(&mut self, i: usize, element: T) -> Result<(), ScriptError> {
         let ri = self.rindex(i)?;
         self.0.insert(ri, element);
         Ok(())
@@ -362,7 +369,7 @@ impl State {
 ///
 /// This is useful for testing & debugging, as we can set up the exact state we want in order to
 /// trigger some behavior.
-pub fn eval_step<'a>(
+fn eval_step<'a>(
     pc: &'a [u8],
     script: &Script,
     flags: VerificationFlags,
@@ -1020,8 +1027,8 @@ pub trait StepFn {
 /// Produces the default stepper, which carries no payload and runs the script as before.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct DefaultStepEvaluator<C> {
-    pub flags: VerificationFlags,
-    pub checker: C,
+    pub(crate) flags: VerificationFlags,
+    pub(crate) checker: C,
 }
 
 impl<C: SignatureChecker + Copy> StepFn for DefaultStepEvaluator<C> {
@@ -1037,7 +1044,7 @@ impl<C: SignatureChecker + Copy> StepFn for DefaultStepEvaluator<C> {
     }
 }
 
-pub fn eval_script<F>(
+fn eval_script<F>(
     stack: Stack<Vec<u8>>,
     script: &Script,
     payload: &mut F::Payload,
@@ -1149,7 +1156,7 @@ where
                 eval_script(remaining_stack, &Script(pub_key_2), payload, stepper)
             })
             .and_then(|p2sh_stack| {
-                if p2sh_stack.last().map_or(false, cast_to_bool) {
+                if p2sh_stack.last().is_ok_and(cast_to_bool) {
                     Ok(p2sh_stack)
                 } else {
                     Err(ScriptError::EvalFalse)
@@ -1160,7 +1167,7 @@ where
     }
 }
 
-pub fn verify_script<F>(
+pub(crate) fn verify_script<F>(
     script_sig: &Script,
     script_pub_key: &Script,
     flags: VerificationFlags,
@@ -1175,7 +1182,7 @@ where
     } else {
         let data_stack = eval_script(Stack::new(), script_sig, payload, stepper)?;
         let pub_key_stack = eval_script(data_stack.clone(), script_pub_key, payload, stepper)?;
-        if pub_key_stack.last().map_or(false, cast_to_bool) {
+        if pub_key_stack.last().is_ok_and(cast_to_bool) {
             if flags.contains(VerificationFlags::P2SH) && script_pub_key.is_pay_to_script_hash() {
                 eval_p2sh(data_stack, script_sig, payload, stepper)
             } else {